MDC rewrite 2.0

+++ src/MacroUtils.h
@@ -5,4 +5,4 @@
 
 #define hang(condition) while (condition) delay(10uL)
 #define avg(x, y) ((x + y) / 2.0)
-#define seconds() (millis() / 1000.0)
+#define seconds() ((double) millis() / 1000.0)
